Vice installation script and various fixes
The nyan vectrex image was not downloaded automatically. On Ubuntu, the
firmware was not included so a script has been provided which should run
on any POSIX compliant shell and any distro (with some modifications).

The installation script has been tested on Ubuntu 16 LTS with dash.

Playing the Grand Prix Circuit example before the music intermezzo
exercise was a mistake as it would only confuse attendees.

The statistics on the floppy slide were wrong and have been fixed.